refactor(test): put mocks closer to their tests (cli)
