+++ test/cli/compileConfig/index.spec.js
@@ -1,30 +1,30 @@
 const path = require("path");
 const expect = require("chai").expect;
 const compileConfig = require("../../../src/cli/compileConfig");
-const fixture = require("../fixtures/rules.sub-not-allowed-error.json");
-const mergedFixture = require("../fixtures/extends/merged.json");
-const mergedArrayOneFixture = require("../fixtures/extends/merged-array-1.json");
-const mergedArrayTwoFixture = require("../fixtures/extends/merged-array-2.json");
+const fixture = require("./mocks/rules.sub-not-allowed-error.json");
+const mergedFixture = require("./mocks/extends/merged.json");
+const mergedArrayOneFixture = require("./mocks/extends/merged-array-1.json");
+const mergedArrayTwoFixture = require("./mocks/extends/merged-array-2.json");
+
+const mockDir = path.join(__dirname, "mocks");
 
 describe("cli/compileConfig", () => {
   it("a rule set without an extends returns just that rule set", () => {
     expect(
-      compileConfig(
-        path.join(__dirname, "../fixtures/rules.sub-not-allowed-error.json")
-      )
+      compileConfig(path.join(mockDir, "rules.sub-not-allowed-error.json"))
     ).to.deep.equal(fixture);
   });
 
   it("a rule set with an extends returns that rule set, extending the mentioned base", () => {
     expect(
-      compileConfig("./test/cli/fixtures/extends/extending.json")
+      compileConfig(path.join(mockDir, "extends/extending.json"))
     ).to.deep.equal(mergedFixture);
   });
 
   it("a rule set with an extends array (0 members) returns that rule set", () => {
     expect(
       compileConfig(
-        "./test/cli/fixtures/extends/extending-array-with-zero-members.json"
+        path.join(mockDir, "extends/extending-array-with-zero-members.json")
       )
     ).to.deep.equal({
       forbidden: [
@@ -40,7 +40,7 @@ describe("cli/compileConfig", () => {
   it("a rule set with an extends array (1 member) returns that rule set, extending the mentioned base", () => {
     expect(
       compileConfig(
-        "./test/cli/fixtures/extends/extending-array-with-one-member.json"
+        path.join(mockDir, "extends/extending-array-with-one-member.json")
       )
     ).to.deep.equal(mergedArrayOneFixture);
   });
@@ -48,7 +48,7 @@ describe("cli/compileConfig", () => {
   it("a rule set with an extends array (>1 member) returns that rule set, extending the mentioned bases", () => {
     expect(
       compileConfig(
-        "./test/cli/fixtures/extends/extending-array-with-two-members.json"
+        path.join(mockDir, "extends/extending-array-with-two-members.json")
       )
     ).to.deep.equal(mergedArrayTwoFixture);
   });
@@ -56,7 +56,7 @@ describe("cli/compileConfig", () => {
   it("a rule set with an extends from node_modules gets merged properly as well", () => {
     expect(
       compileConfig(
-        "./test/cli/fixtures/extends/extending-from-node-modules.json"
+        path.join(mockDir, "extends/extending-from-node-modules.json")
       )
     ).to.deep.equal({
       allowed: [
@@ -79,17 +79,15 @@ describe("cli/compileConfig", () => {
 
   it("borks on a circular extends (1 step)", () => {
     const lMessageOutTake = `config is circular - ${path.join(
-      __dirname,
-      "../fixtures/extends/circular-one.js"
-    )} -> ${path.join(
-      __dirname,
-      "../fixtures/extends/circular-two.js"
-    )} -> ${path.join(__dirname, "../fixtures/extends/circular-one.js")}.`;
+      mockDir,
+      "extends/circular-one.js"
+    )} -> ${path.join(mockDir, "extends/circular-two.js")} -> ${path.join(
+      mockDir,
+      "extends/circular-one.js"
+    )}.`;
 
     expect(() => {
-      compileConfig(
-        path.join(__dirname, "../fixtures/extends/circular-one.js")
-      );
+      compileConfig(path.join(mockDir, "extends/circular-one.js"));
     }).to.throw(lMessageOutTake);
   });
 });
