+++ test/cli/compileConfig/mocks/extends/base-for-base.js
