+++ test/cli/compileConfig/mocks/extends/base-for-extends.js
