+++ test/cli/compileConfig/mocks/extends/circular-one.js
