+++ test/cli/compileConfig/mocks/extends/circular-two.js
