+++ test/cli/compileConfig/mocks/extends/extending-array-with-one-member.json
