+++ test/cli/compileConfig/mocks/extends/extending-array-with-two-members.json
