+++ test/cli/compileConfig/mocks/extends/extending-array-with-zero-members.json
