+++ test/cli/compileConfig/mocks/extends/extending-from-node-modules.json
