+++ test/cli/compileConfig/mocks/extends/extending.json
