+++ test/cli/compileConfig/mocks/extends/merged-array-1.json
