+++ test/cli/compileConfig/mocks/extends/merged-array-2.json
