+++ test/cli/compileConfig/mocks/extends/merged.json
