+++ test/cli/compileConfig/mocks/extends/second-base-for-extends.json
